add float

+++ src/Calculadora.java
@@ -1,12 +1,12 @@
 public class Calculadora {
-    public int somar (int a, int b){
+    public float somar (float a, float b){
         return  a+b;
     }
 
-    public int subtrai (int a, int b){
+    public float subtrai (float a, float b){
         return  a-b;
     }
-    public int multi (int a, int b){
+    public float multi (float a, float b){
         return  a*b;
     }
     public float div (float a, float b){
